requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
google-generativeai==0.3.1
twilio==8.11.0
python-dotenv==1.0.0
